Added SiteImage to Repository pattern, added upsert, added encoding for site image. display image from database

+++ EyonSolution/Eyon.DataAccess/Data/Repository/IRepository/IRepository.cs
@@ -7,15 +7,16 @@ using System.Text;
 namespace Eyon.DataAccess.Data.Repository.IRepository
 {
     public interface IRepository<T> where T : class
-    {
-        T Get(int id);
+    {        
+        T Get(long id);
         IEnumerable<T> GetAll(Expression<Func<T, bool>> filter = null, Func<IQueryable<T>,
                         IOrderedQueryable<T>> orderBy = null, string includeProperties = null);
 
         T GetFirstOrDefault(Expression<Func<T, bool>> filter = null, string includeProperties = null);
 
-        void Add(T entity);
-        void Remove(int id);
+        void Add(T entity);        
+
+        void Remove(long id);
         void Remove(T entity);
     }
 }
